🚀 feat(PreferenceSwitcher.tsx): add PreferenceSwitcher component to toggle bookmark favicon visibility and color mode 🔨 refactor(Settings.tsx): import PreferenceSwitcher from correct file 🔨 refactor(Main.tsx): move getConfig() into InnerComponent to hot reload configs and use it in Main component
